Add readtxt() for reading in text files zipped or plain.
This is a preparation for wrapping the ImageJ macro templates into a .zip file
so they don't show up as available macros in the Fiji plugin menu anymore
(which they clearly aren't).

This probably wouldn't be required if Fiji's Jython was able to import classes
from files within a .jar, which dosn't work. As of now it works for "plain"
modules but not for objects.