Always return list(str) from gen_tile_config().
Otherwise writing the tile config will fail as writelines() cannot handle empty
lists. In our case it is better to still write a tile config that contains an
error message in a comment as this will not prevent the processing of the other
mosaics and it can easily be investigated when a user reports a problem.