Fix control flow on broken subvolumes.
The previous implementation did not stop processing the subvolumes if one of
them was broken, but continued with the next one of the dataset. This was never
an issue until we came across an experiment where one mosaic had some missing
subvolumes, but later ones were present. In that case the code was trying to
add them as subvolume, whereas the dataset itself had been set to "None"
already (which caused an exception then).